feat: implement core trajectory state machines and domain-specific transition logic

+++ coh-node/crates/coh-core/src/external/adapters.rs
@@ -22,7 +22,7 @@ fn valid_signature(step_index: u64, signer: &str) -> SignatureWire {
     SignatureWire {
         signature: format!("sig-{:016}", step_index),
         signer: signer.to_string(),
-        timestamp: 1_700_000_000 + step_index as u64,
+        timestamp: 1_700_000_000 + step_index,
     }
 }
